% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crrry-package.R
\docType{package}
\name{crrry-package}
\alias{crrry}
\alias{crrry-package}
\title{crrry: 'crrri' recipes for 'shiny'}
\description{
More about what it does (maybe more than one line)
    Use four spaces when indenting paragraphs within the Description.
}
\author{
\strong{Maintainer}: Colin Fay \email{colin@thinkr.fr}

}
\keyword{internal}
