% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch.R
\name{CrrryGeneric}
\alias{CrrryGeneric}
\title{Generic crrry methods}
\description{
Generic crrry methods

Generic crrry methods
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{process}}{Chrome process, frm \code{{crrri}}}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-call_js}{\code{CrrryGeneric$call_js()}}
\item \href{#method-click_on_id}{\code{CrrryGeneric$click_on_id()}}
\item \href{#method-gremlins_horde}{\code{CrrryGeneric$gremlins_horde()}}
\item \href{#method-shiny_set_input}{\code{CrrryGeneric$shiny_set_input()}}
\item \href{#method-wait_for_shiny_ready}{\code{CrrryGeneric$wait_for_shiny_ready()}}
\item \href{#method-wait_for}{\code{CrrryGeneric$wait_for()}}
\item \href{#method-clone}{\code{CrrryGeneric$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-call_js"></a>}}
\if{latex}{\out{\hypertarget{method-call_js}{}}}
\subsection{Method \code{call_js()}}{
Execute JavaScript code in the session.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CrrryGeneric$call_js(code, check = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{code}}{JS code}

\item{\code{check}}{Should \code{{crrry}} check if Shiny is still running?}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-click_on_id"></a>}}
\if{latex}{\out{\hypertarget{method-click_on_id}{}}}
\subsection{Method \code{click_on_id()}}{
Click on an id
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CrrryGeneric$click_on_id(id, check = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{ID}

\item{\code{check}}{Should \code{{crrry}} check if Shiny is still running?}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gremlins_horde"></a>}}
\if{latex}{\out{\hypertarget{method-gremlins_horde}{}}}
\subsection{Method \code{gremlins_horde()}}{
Send a \code{gremlin.js} horde
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CrrryGeneric$gremlins_horde(check = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{check}}{Should \code{{crrry}} check if Shiny is still running?}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-shiny_set_input"></a>}}
\if{latex}{\out{\hypertarget{method-shiny_set_input}{}}}
\subsection{Method \code{shiny_set_input()}}{
Set the value of a shiny input
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CrrryGeneric$shiny_set_input(id, val, check = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{Shiny ID}

\item{\code{val}}{Value for the ID}

\item{\code{check}}{Should \code{{crrry}} check if Shiny is still running?}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-wait_for_shiny_ready"></a>}}
\if{latex}{\out{\hypertarget{method-wait_for_shiny_ready}{}}}
\subsection{Method \code{wait_for_shiny_ready()}}{
Wait for Shiny to be ready
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CrrryGeneric$wait_for_shiny_ready(check = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{check}}{Should \code{{crrry}} check if Shiny is still running?}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-wait_for"></a>}}
\if{latex}{\out{\hypertarget{method-wait_for}{}}}
\subsection{Method \code{wait_for()}}{
Wait for a JS condition to be TRUE
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CrrryGeneric$wait_for(cond, check = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{cond}}{JS condition}

\item{\code{check}}{Should \code{{crrry}} check if Shiny is still running?}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CrrryGeneric$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
