% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch.R
\name{CrrryProc}
\alias{CrrryProc}
\title{Launch a crrrry on a local process}
\value{
A crrrry object
}
\description{
Launch a crrrry on a local process

Launch a crrrry on a local process
}
\section{Super class}{
\code{\link[crrry:CrrryGeneric]{crrry::CrrryGeneric}} -> \code{CrrryOnPage}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{CrrryProc$new()}}
\item \href{#method-stop}{\code{CrrryProc$stop()}}
\item \href{#method-is_alive}{\code{CrrryProc$is_alive()}}
\item \href{#method-clone}{\code{CrrryProc$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="crrry" data-topic="CrrryGeneric" data-id="call_js">}\href{../../crrry/html/CrrryGeneric.html#method-call_js}{\code{crrry::CrrryGeneric$call_js()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="crrry" data-topic="CrrryGeneric" data-id="click_on_id">}\href{../../crrry/html/CrrryGeneric.html#method-click_on_id}{\code{crrry::CrrryGeneric$click_on_id()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="crrry" data-topic="CrrryGeneric" data-id="gremlins_horde">}\href{../../crrry/html/CrrryGeneric.html#method-gremlins_horde}{\code{crrry::CrrryGeneric$gremlins_horde()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="crrry" data-topic="CrrryGeneric" data-id="shiny_set_input">}\href{../../crrry/html/CrrryGeneric.html#method-shiny_set_input}{\code{crrry::CrrryGeneric$shiny_set_input()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="crrry" data-topic="CrrryGeneric" data-id="wait_for">}\href{../../crrry/html/CrrryGeneric.html#method-wait_for}{\code{crrry::CrrryGeneric$wait_for()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="crrry" data-topic="CrrryGeneric" data-id="wait_for_shiny_ready">}\href{../../crrry/html/CrrryGeneric.html#method-wait_for_shiny_ready}{\code{crrry::CrrryGeneric$wait_for_shiny_ready()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Wait for a JS condition to be TRUE
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CrrryProc$new(
  chrome_bin = Sys.getenv("HEADLESS_CHROME"),
  fun = "pkgload::load_all();run_app()",
  shiny_port = 2811L,
  chrome_port = 9222L,
  headless = TRUE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{chrome_bin}}{Path to Chrome binary, passed to \code{Chrome$new()}}

\item{\code{fun}}{A function launching the shiny app}

\item{\code{shiny_port}}{The port to launch the shiny apps on}

\item{\code{chrome_port}}{Chrome_port, passed to \code{Chrome$new()}}

\item{\code{headless}}{Run headless? Passed to \code{Chrome$new()}}

\item{\code{...}}{Futher args passed to \code{Chrome$new()}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-stop"></a>}}
\if{latex}{\out{\hypertarget{method-stop}{}}}
\subsection{Method \code{stop()}}{
Stop the process
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CrrryProc$stop()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-is_alive"></a>}}
\if{latex}{\out{\hypertarget{method-is_alive}{}}}
\subsection{Method \code{is_alive()}}{
Check if the url is available
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CrrryProc$is_alive()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CrrryProc$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
